Fixed demon paths so that Double-tall grass doesn't mess it up

+++ src/main/java/WayofTime/alchemicalWizardry/common/spell/complex/effect/SpellHelper.java
@@ -1,12 +1,10 @@
 package WayofTime.alchemicalWizardry.common.spell.complex.effect;
 
-import WayofTime.alchemicalWizardry.AlchemicalWizardry;
-import WayofTime.alchemicalWizardry.api.alchemy.energy.IAlchemyGoggles;
-import WayofTime.alchemicalWizardry.api.items.interfaces.IReagentManipulator;
-import WayofTime.alchemicalWizardry.api.soulNetwork.SoulNetworkHandler;
-import WayofTime.alchemicalWizardry.api.spell.APISpellHelper;
-import WayofTime.alchemicalWizardry.common.NewPacketHandler;
-import cpw.mods.fml.common.FMLCommonHandler;
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Random;
+
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockLiquid;
 import net.minecraft.entity.Entity;
@@ -37,11 +35,13 @@ import net.minecraft.world.WorldServer;
 import net.minecraftforge.common.util.FakePlayer;
 import net.minecraftforge.common.util.ForgeDirection;
 import net.minecraftforge.oredict.OreDictionary;
-
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Random;
+import WayofTime.alchemicalWizardry.AlchemicalWizardry;
+import WayofTime.alchemicalWizardry.api.alchemy.energy.IAlchemyGoggles;
+import WayofTime.alchemicalWizardry.api.items.interfaces.IReagentManipulator;
+import WayofTime.alchemicalWizardry.api.soulNetwork.SoulNetworkHandler;
+import WayofTime.alchemicalWizardry.api.spell.APISpellHelper;
+import WayofTime.alchemicalWizardry.common.NewPacketHandler;
+import cpw.mods.fml.common.FMLCommonHandler;
 
 public class SpellHelper
 {
@@ -503,7 +503,7 @@ public class SpellHelper
         return false;
     }
 
-    public static Entity teleportEntitySameDim(int x, int y, int z, Entity entity)
+    public static Entity teleportEntitySameDim(double x, double y, double z, Entity entity)
     {
         if (entity != null)
         {
